+++ CHANGELOG.md
@@ -8,6 +8,7 @@ will be documented here.
 - Added language server for use in vscode extension
 - Implemented primitives types and operations for Solana
 - Functions can be declared outside of contracts
+- Constants can be declared outside of contracts
 
 ## [0.1.5]